Bug 928537 - Datachannel streams should be kept in the range 1 to MAX_NUM_STREAMS r=jesup
UltraBlame original commit: 1a75117d380fc6146067ea75d04a776c7de3aacf

+++ media/webrtc/signaling/src/sipcc/core/sdp/sdp.h
@@ -714,7 +714,7 @@ typedef struct sdp_fmtp {
 
 typedef struct sdp_sctpmap {
     u16                       port;
-    u16                       streams;   
+    u32                       streams;   
     char                      protocol[SDP_MAX_STRING_LEN+1];
 } sdp_sctpmap_t;
